test(http): Extract cookie tests from `ApplicationTest` to `ApplicationCookieTest` class.
